# openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise E-Commerce API
  version: 0.1.0
  description: REST API for catalog, checkout, orders, and payments.
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /api/v1/admin/orders:
    get:
      summary: Admin list orders
      operationId: adminListOrders
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [confirmed, shipped, canceled]
        - name: export_pending
          in: query
          description: Confirmed orders not yet exported to MoySklad
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Paginated order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderListResponse"
  /api/v1/admin/orders/{order_number}:
    get:
      summary: Admin get order detail
      operationId: adminGetOrder
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: order_number
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail with lines and status history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetailSchema"
        "404":
          description: Order not found
  /api/v1/admin/orders/{order_number}/status:
    patch:
      summary: Admin update order status (ship or cancel)
      operationId: adminUpdateOrderStatus
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: order_number
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetailSchema"
        "404":
          description: Order not found
        "422":
          description: Invalid status transition
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderLine:
      type: object
      required: [id, variant_id, quantity, unit_price_cents, line_total_cents, product_snapshot]
      properties:
        id:
          type: string
          format: uuid
        variant_id:
          type: string
          format: uuid
        quantity:
          type: integer
        unit_price_cents:
          type: integer
        line_total_cents:
          type: integer
        product_snapshot:
          type: object
          description: Immutable product/variant snapshot at line creation time
          properties:
            variant_id:
              type: string
              format: uuid
            sku:
              type: string
            name:
              type: string
            product_id:
              type: string
              format: uuid
            product_name:
              type: string
            product_slug:
              type: string
            price_cents:
              type: integer
              description: Retail reference price at snapshot time
            currency:
              type: string
            price_tier:
              type: string
              enum: [retail, wholesale]
              description: Commercial tier applied to unit_price_cents (ADR-008)
...
    AdminOrderSummarySchema:
      type: object
      required:
        - id
        - order_number
        - status
        - currency
        - total_cents
        - created_at
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        status:
          type: string
          enum: [confirmed, shipped, canceled]
        currency:
          type: string
        total_cents:
          type: integer
        customer_email:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AdminOrderStatusHistorySchema:
      type: object
      required: [id, to_status, changed_by, changed_at]
      properties:
        id:
          type: string
          format: uuid
        from_status:
          type: string
          nullable: true
        to_status:
          type: string
        changed_by:
          type: string
        reason:
          type: string
          nullable: true
        changed_at:
          type: string
          format: date-time
    AdminOrderDetailSchema:
      type: object
      required:
        - id
        - order_number
        - status
        - currency
        - subtotal_cents
        - discount_cents
        - shipping_cents
        - tax_cents
        - total_cents
        - created_at
        - updated_at
        - lines
        - status_history
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        status:
          type: string
          enum: [confirmed, shipped, canceled]
        currency:
          type: string
        subtotal_cents:
          type: integer
        discount_cents:
          type: integer
        shipping_cents:
          type: integer
        tax_cents:
          type: integer
        total_cents:
          type: integer
        customer_email:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
        status_history:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderStatusHistorySchema"
    AdminOrderListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderSummarySchema"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [shipped, canceled]
        reason:
          type: string
          maxLength: 500
...